# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FeedKeeper native API
  version: "1"
  license: { name: MIT, url: "https://github.com/visualfusion/feedkeeper/blob/main/LICENSE" }
  description: |
    The API for native apps, served under `/api/v1`. Design notes and what is still planned
    are in `docs/design/native-api.md`. A test keeps this file in step with the implemented routes.

    Authenticate with `Authorization: Bearer <token>`: a device token from pairing, or a personal
    access token. Read-only tokens can call every `GET` and nothing that changes data.
servers:
  - url: /api/v1
security:
  - bearer: []
...
paths:
...
  /pairing-codes:
    post:
      tags: [Devices]
      summary: Create a one-time pairing code
      description: Called by the web app with the user's session. The code is valid for 10 minutes and replaces any earlier one.
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/PairingCode" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /devices/pair:
    post:
      tags: [Devices]
      summary: Exchange a pairing code for a device token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, deviceName, platform]
              properties:
                code: { type: string, description: "As shown, or typed without dashes and in any case" }
                deviceName: { type: string, maxLength: 100 }
                platform: { type: string, maxLength: 40, examples: [ios] }
                appVersion: { type: string, maxLength: 40 }
      responses:
        "201":
          description: Paired. The token is shown only here.
          content:
            application/json:
              schema:
                type: object
                properties:
                  deviceId: { type: integer }
                  token: { type: string }
                  user: { type: object, properties: { id: { type: integer }, email: { type: string }, displayName: { type: string } } }
        "400": { $ref: "#/components/responses/Invalid" }
        "429": { description: Too many attempts }
  /me:
    get:
      tags: [Devices]
      summary: The signed-in user
      description: |
        Uncached account profile. Technical support from /meta, account capabilities and
        token scope are independent. Missing capabilities on an older server mean legacy
        compatibility; missing keys in an existing feature map deny access.
        See docs/design/native-api.md for operation mappings and downgrade behavior.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [id, email, displayName, role, scope, deviceId, capabilities]
                properties:
                  id: { type: integer }
                  email: { type: string }
                  displayName: { type: string }
                  role: { type: string, enum: [admin, user] }
                  scope: { type: string, enum: [read, write] }
                  deviceId: { type: [integer, "null"] }
                  capabilities: { $ref: "#/components/schemas/AccountCapabilities" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /devices/{id}:
    delete:
      tags: [Devices]
      summary: Revoke a device
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/NotFound" }
  /sync:
    get:
      tags: [Sync]
      summary: Changes to the user's own data since a position
      description: |
        The log holds one entry per object, so `since=0` returns the complete state. A client whose
        position is older than a forgotten deletion gets `410 resync_required` and starts over from 0.
      parameters:
        - { name: since, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 200 } }
      responses:
        "403": { $ref: "#/components/responses/CapabilityDenied" }
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/SyncPage" } } } }
        "410": { description: resync_required }
  /mutations:
    post:
      tags: [Sync]
      summary: Apply changes made offline
      description: Each mutation has a client-made id, so a retried request changes nothing. Read, saved and reading position are last-writer-wins per field.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mutations]
              properties:
                mutations: { type: array, minItems: 1, maxItems: 200, items: { $ref: "#/components/schemas/Mutation" } }
      responses:
        "200":
          description: One result per mutation, in order
          content:
            application/json:
              schema:
                type: object
                properties:
                  results: { type: array, items: { $ref: "#/components/schemas/MutationResult" } }
                  seq: { type: integer, description: Current end of the change log }
        "400": { $ref: "#/components/responses/Invalid" }
        "403": { $ref: "#/components/responses/ReadOnly" }
...
  /items:
    get:
      tags: [Articles]
      summary: Articles, newest added first, with a cursor
      description: |
        Compact by default. Use `nextCursor` as `before` for older pages and `newestCursor` as `after`
        for newly added articles. With `include=content` the page is limited to 50 articles.
      parameters:
        - { name: subscriptionId, in: query, schema: { type: integer } }
        - { name: folderId, in: query, schema: { type: integer } }
        - { name: unread, in: query, schema: { type: boolean } }
        - { name: saved, in: query, schema: { type: boolean } }
        - { name: q, in: query, description: "Searches titles and summaries, and the full text of saved articles", schema: { type: string, maxLength: 200 } }
        - { name: since, in: query, schema: { type: string }, description: Published at or after this ISO date or date-time }
        - { name: until, in: query, schema: { type: string }, description: Published at or before this ISO date or date-time }
        - { name: after, in: query, schema: { type: string } }
        - { name: before, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { name: include, in: query, schema: { type: string, enum: [content] } }
        - { $ref: "#/components/parameters/Images" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Item" } }
                  newestCursor: { type: [string, "null"] }
                  nextCursor: { type: [string, "null"] }
        "400": { $ref: "#/components/responses/Invalid" }
...
  /items/{id}:
    get:
      tags: [Articles]
      summary: One article with its content
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { $ref: "#/components/parameters/Images" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Item" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /items/{id}/image:
    get:
      tags: [Images]
      summary: An image the article shows, through this server
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { name: src, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: A raster image, content: { "image/*": { schema: { type: string, format: binary } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "415": { description: Not a supported image }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer } }
    Images:
      name: images
      in: query
      description: "`proxy` points the images of articles that are not archived at this server, so a client can keep them offline"
      schema: { type: string, enum: [original, proxy] }
  responses:
    Unauthorized: { description: Missing or invalid credentials }
    NotFound: { description: "Not found, or not visible to this user" }
    Invalid: { description: "Invalid input: { error: invalid_input }" }
    ReadOnly: { description: The token is read-only }
    CapabilityDenied:
      description: "The account lacks the capability or the token is read-only; stored content is retained"
      content:
        application/json:
          schema:
            oneOf:
              - { $ref: "#/components/schemas/CapabilityDenial" }
              - type: object
                required: [error]
                properties:
                  error: { type: string, const: read_only_token }
  schemas:
    AccountCapabilities:
      type: object
      required: [type, features, manageUrl]
      properties:
        type: { type: string, example: selfhosted, description: "Opaque display identifier; never overrides feature permissions" }
        features:
          type: object
          description: "Known keys: mcp, sync, notes, editions, fulltext, search.fts. Explicit false and missing keys deny access. Self-hosted instances enable all known keys."
          additionalProperties: { type: boolean }
        manageUrl: { type: [string, "null"], description: "Optional HTTPS account-management URL; does not grant an app license" }
    CapabilityDenial:
      type: object
      required: [error, capability, manageUrl]
      properties:
        error: { type: string, const: capability_not_available }
        capability: { type: string }
        manageUrl: { type: [string, "null"] }
...
    PairingCode:
      type: object
      properties:
        code: { type: string }
        expiresAt: { type: string, format: date-time }
        pairingUrl: { type: string, description: "The content of the QR code" }
...
    ItemState:
      type: object
      properties:
        read: { type: boolean }
        saved: { type: boolean }
        savedAt: { type: [string, "null"] }
        archivedAt: { type: [string, "null"] }
        progress: { type: [number, "null"], minimum: 0, maximum: 1 }
        hasNote: { type: boolean }
    ArticleContent:
      type: object
      required: [revision, source, status, attemptedAt, retryAt]
      properties:
        revision: { type: integer, minimum: 1, description: Increases when publisher content or article identity/image changes }
        source: { type: string, enum: [reader, feed, snippet] }
        status: { type: string, enum: [pending, ready, feed_only, consent_wall, paywall, bot_blocked, timeout, partial, failed] }
        attemptedAt: { type: [string, "null"], format: date-time }
        retryAt: { type: [string, "null"], format: date-time }
    Item:
      type: object
      properties:
        id: { type: integer }
        subscriptionId: { type: integer }
        subscriptionTitle: { type: [string, "null"] }
        title: { type: [string, "null"] }
        url: { type: [string, "null"] }
        publishedAt: { type: [string, "null"] }
        addedAt: { type: string }
        snippet: { type: [string, "null"] }
        imageUrl: { type: [string, "null"] }
        hasFullText: { type: boolean }
        content: { $ref: "#/components/schemas/ArticleContent" }
        contentHtml: { type: [string, "null"], description: Only with content included }
        fullTextHtml: { type: [string, "null"], description: Only with content included }
        state: { $ref: "#/components/schemas/ItemState" }
...
    NoteConflictState:
      type: object
      required: [itemId, content, revision, createdAt, updatedAt, deleted]
      properties:
        itemId: { type: integer }
        content: { type: [string, "null"], description: "Null after deletion" }
        revision: { type: integer, minimum: 0, description: "Increases on every edit, deletion and recreation" }
        createdAt: { type: [string, "null"], format: date-time }
        updatedAt: { type: [string, "null"], format: date-time }
        deleted: { type: boolean }
...
    SyncPage:
      type: object
      properties:
        changes: { type: array, items: { $ref: "#/components/schemas/Change" } }
        nextSeq: { type: integer, description: Pass as `since` next time }
        hasMore: { type: boolean }
    Change:
      type: object
      properties:
        seq: { type: integer }
        entity: { type: string, enum: [item_state, item_content, subscription, folder, muted_keyword, note, edition, native_preferences] }
        id: { type: integer }
        op: { type: string, enum: [upsert, delete] }
        data: { type: object, description: "Current data for upserts. Note deletions carry {revision}; other deletions omit data." }
    Mutation:
      type: object
      required: [id, type, itemId, at]
      properties:
        id: { type: string, minLength: 8, maxLength: 64, description: Made by the client; makes retries harmless }
        type: { type: string, enum: [item.read, item.save, item.progress, item.note.set, item.note.delete] }
        itemId: { type: integer }
        value: { type: boolean, description: For item.read and item.save }
        position: { type: number, minimum: 0, maximum: 1, description: "For item.progress; below 5 % or above 95 % clears it" }
        content: { type: string, description: For item.note.set }
        expectedRevision: { type: integer, minimum: 0, description: "For item.note.set and item.note.delete; omit only for unconditional changes" }
        at: { type: string, format: date-time, description: "When the user did it; at most 5 minutes ahead of the server, at most 30 days old" }
    MutationResult:
      description: "A capability denial is retryable and does not consume the mutation UUID. Other entries in the batch still apply."
      type: object
      properties:
        id: { type: string }
        outcome: { type: string, enum: [applied, duplicate, stale, rejected, conflict] }
        error: { type: string, description: "invalid_mutation, invalid_time, item_not_found, revision_conflict, or capability_not_available" }
        capability: { type: string, description: "Required account capability when access was denied" }
        manageUrl: { type: [string, "null"] }
        retryable: { type: boolean, description: "True for capability denials; retain the mutation for a later retry" }
        current: { $ref: "#/components/schemas/NoteConflictState" }